Mejoras

+++ app/Http/Controllers/AccionMejoraController.php
@@ -2,9 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use DateTime;
 use App\Models\AccionMejora;
 use App\Models\ActividadAccion;
 use App\Models\User;
+use App\Models\Resultado;
+use App\Models\Indicador;
 use App\Models\encargado_accion;
 
 use App\Models\estandar;
@@ -33,7 +36,9 @@ class AccionMejoraController extends Controller
     public function create($idPlan)
     {
         $users = User::all();
-        return view('accionmejora.create',['idPlan'=>$idPlan],compact(['users']));
+        $resultados = Resultado::all();
+        $indicadores = Indicador::all();
+        return view('accionmejora.create',['idPlan'=>$idPlan],compact(['users','resultados','indicadores']));
     }
 
     /**
@@ -51,7 +56,7 @@ class AccionMejoraController extends Controller
             'valor' => 'required|integer',
             'fechaInicio' => 'required|date|max:255',
             'fechaFin' => 'required|date|max:255',
-            'duracion' => 'required|integer',
+            //'duracion' => 'required|integer',
             'semestreEjecucion' => 'required|string|max:255',
             'recursos' => 'required|string|max:255',
             'metas' => 'required|string|max:255',
@@ -67,6 +72,19 @@ class AccionMejoraController extends Controller
         ];
         $this->validate($request,$campos,$mensaje);
         $datosAccion = request()->except('_token');
+
+
+        $fdate = $request->fechaInicio;
+        $tdate = $request->fechaFin;
+        $datetime1 = new DateTime($fdate);
+        $datetime2 = new DateTime($tdate);
+        $interval = $datetime1->diff($datetime2);
+        $daysDiff = $interval->format('%a');//now do whatever you like with $days
+        $semanas = floor($daysDiff / 7);
+        $days = floor($daysDiff % 7);
+        $duracion = $semanas.' semana(s)';
+        if($days>0){ $duracion = $duracion.' con '.$days.' dias.'; }
+        $datosAccion['duracion']= $duracion;
         AccionMejora::insert($datosAccion);
         return redirect('planes/' . $datosAccion['idPlan'] . '/edit')->with('mensaje','Acción de mejora creada');
     }
@@ -94,12 +112,12 @@ class AccionMejoraController extends Controller
         $actividades = ActividadAccion::where('idAccion','=',$accion->id)->paginate(10);
         $users = User::all();
         $encargados = encargado_accion::where('idAccion','=',$accion->id)->get();
-
-        
         $estandares = estandar::all();
+        $resultados = Resultado::all();
+        $indicadores = Indicador::all();
         $estandares_accion = estandar_accion::where('idAccion','=',$accion->id)->get();
 
-        return view('accionmejora.edit',compact(['accion','actividades','users','encargados','estandares','estandares_accion']));
+        return view('accionmejora.edit',compact(['accion','actividades','users','encargados','estandares','estandares_accion','resultados','indicadores']));
     }
 
     /**
@@ -112,26 +130,43 @@ class AccionMejoraController extends Controller
     public function update(Request $request, $id)
     {
         $datosAccion = request()->except(['_token','_method','encargados','estandares']);
+        
+        $fdate = $request->fechaInicio;
+        $tdate = $request->fechaFin;
+        $datetime1 = new DateTime($fdate);
+        $datetime2 = new DateTime($tdate);
+        $interval = $datetime1->diff($datetime2);
+        $daysDiff = $interval->format('%a');//now do whatever you like with $days
+        $semanas = floor($daysDiff / 7);
+        $days = floor($daysDiff % 7);
+        $duracion = $semanas.' semana(s)';
+        if($days>0){ $duracion = $duracion.' con '.$days.' dias.'; }
+        $datosAccion['duracion']= $duracion;
+
         AccionMejora::where('id','=',$id)->update($datosAccion);
 
         $encargados = $request->input('encargados');
-        encargado_accion::where('idAccion',$id)->delete();
-        for ($i=0; $i < count($encargados); $i++) {
-            $datoEncargado[] = [
-                'idAccion'  => $id,
-                'idUsuario' => $encargados[$i]
-            ];
-            encargado_accion::insert($datoEncargado);
+        if($encargados){
+            encargado_accion::where('idAccion',$id)->delete();
+            for ($i=0; $i < count($encargados); $i++) {
+                $datoEncargado[] = [
+                    'idAccion'  => $id,
+                    'idUsuario' => $encargados[$i]
+                ];
+                encargado_accion::insert($datoEncargado);
+            }
         }
 
         $estandares = $request->input('estandares');
-        estandar_accion::where('idAccion',$id)->delete();
-        for ($i=0; $i < count($estandares); $i++) {
-            $datoEstandares[] = [
-                'idAccion'  => $id,
-                'idEstandar' => $estandares[$i]
-            ];
-            estandar_accion::insert($datoEstandares);
+        if($estandares){
+            estandar_accion::where('idAccion',$id)->delete();
+            for ($i=0; $i < count($estandares); $i++) {
+                $datoEstandares[] = [
+                    'idAccion'  => $id,
+                    'idEstandar' => $estandares[$i]
+                ];
+                estandar_accion::insert($datoEstandares);
+            }
         }
 
         return redirect('planes/' . $datosAccion['idPlan'] . '/edit')->with('mensaje','Accion de mejora modificada');
